Sorted out OAuth error and "valid instance config" attr
Sorted out OAuth error when OAuth was actually successful. In short, web browser was making a call for the "favicon" which didn't have a query paramter, so we just through an error (Cause bad checking). We now look for favicon...

Also, added a new attribute to the SNICHConfig.Instance interface called config_status so that we know if the instance had been completed in it's configuration. We then use that to show instances when "performing additional comands". In short, no need to show an instance if it's config is incpomplete or it was aborted.